requirements.txt
# Core dependencies
semantic-kernel>=1.22.0
python-dotenv>=1.0.1

# AI and Azure integrations
azure-monitor-opentelemetry-exporter>=1.0.0b39  # Include pre-release versions
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation>=0.41b0

# Data validation and type checking
pydantic>=2.0.0
typing-extensions>=4.0.0

# Code execution sandbox
ai-code-sandbox>=0.1.0

# HTTP client
aiohttp>=3.10.5

# Optional: chainlit for UI (if needed)
chainlit>=2.0.1

# Search and Web APIs
tavily-python

# optional: pytest
pytest
